shared 분리
